requirements.txt
boto==2.38.0
django-allauth==0.24.1
Django==1.8.7
django-bower==5.0.4
django-cleanup==0.4.2
django-ganalytics==0.4
django-multiupload==0.5
django-ses==0.7.0
django-storages-redux==1.3
django.js==0.8.1
gpxpy==1.0.0
gunicorn==19.4.1
lxml==3.5.0
newrelic==2.60.0.46
numpy==1.10.2
Pint==0.6.0
python-dateutil==2.4.2
pytz==2015.7
